Fade out function set up
Removed unnecessary scripts, main menu fades out on start
