--- Assets/Scripts/Managers/ClanLeaderManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fd310c9605a5a2d4692ba3b1f03dce1f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
