+++ Assets/Scripts/Managers/EndScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b8a0cc1a6e44e842926a410d2456f42
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
