+++ Assets/Scripts/Managers/FadeManager.cs
@@ -1,0 +1,43 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class FadeManager : MonoBehaviour
+{
+    public static FadeManager Instance {get; private set;}
+
+    [Header("Fade Settings")]
+    public Image fadeImage;
+    public float fadeDuration;
+
+    private void Start()
+    {
+        if (fadeImage != null)
+        {
+            fadeImage.gameObject.SetActive(false);
+        }
+    }
+
+    public IEnumerator FadeOut()
+    {
+        if (fadeImage != null)
+        {
+            fadeImage.gameObject.SetActive(true);
+            Color fadeColor = fadeImage.color;
+            float timer = 0f;
+
+            while (timer < fadeDuration)
+            {
+                timer += Time.deltaTime;
+                fadeColor.a = Mathf.Lerp(0f, 1f, timer / fadeDuration);
+                fadeImage.color = fadeColor;
+                yield return null;
+            }
+
+            fadeColor.a = 1f;
+            fadeImage.color = fadeColor;
+        }
+    }
+
+}
