+++ Assets/Scripts/Managers/FadeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 326bf9839af49b24aa232045e619f0bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
