--- Assets/Scripts/Managers/FarmManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4dd60c1ec50fe5b468140bc7751a4cad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
