--- Assets/Scripts/Managers/FishermanManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f76509d84efb14841bde460030a50eb1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
