--- Assets/Scripts/Managers/HouseManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 80715353b2c1ba9469b3526b198b879d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
