--- Assets/Scripts/Managers/RiverManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 50efa4eeb68584a40b501fb73cdecaf6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
