--- Assets/Scripts/Managers/VillageManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 002818cafbea6d548928e8154b2ea4f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
