--- Assets/Scripts/PlayerInputManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 70d4f630cfa196949b7fc42955c26670
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
