--- Assets/Scripts/SecretTree.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b64c8b654f2cca42af86620e97844da
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
